requirement.txt
# ML
scikit-surprise==1.1.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2

# Backend
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
cachetools==5.3.2
python-multipart==0.0.6

# Utils
scipy==1.11.4